requirements_test.txt
accelerate==1.0.1
datasets==3.1.0
decord==0.6.0
deepspeed==0.9.5
diffusers==0.30.0
einops==0.6.1
ftfy==6.2.3
gradio==3.37.0
imageio==2.35.1
iopath==0.1.10
joblib==1.4.2
matplotlib==3.7.5
numpy==1.24.1
pandas==2.0.3
peft==0.4.0
Pillow==10.2.0
pytorchvideo==0.1.5
PyYAML==6.0.2
regex==2024.11.6
scipy==1.10.1
timm==0.6.13
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
transformers==4.46.3
tensorboard==2.17.1
dgl==1.1.2
